+++ daemons/controld/controld_execd_state.c
@@ -78,7 +78,7 @@ fail_pending_op(gpointer key, gpointer value, gpointer user_data)
     event.interval_ms = op->interval_ms;
     lrmd__set_result(&event, PCMK_OCF_UNKNOWN_ERROR, PCMK_EXEC_NOT_CONNECTED,
                      "Action was pending when executor connection was dropped");
-    event.t_run = (unsigned int) op->start_time;
+    event.t_run = op->start_time;
     event.t_rcchange = (unsigned int) op->start_time;
 
     event.call_id = op->call_id;
